Updated es_MX.lang file for BoB (first 100 lines only)
The first 100 lines for this Mexican Spanish translation .lang file have been fully retranslated and revised. The goal is to fix all the translations and send the progress every 100 lines translated


P.S. If you're Poggy, could you please check DMs? It's to keep you progress live on the translations